feat: add install_requires in setup file and move odpf branding to raystack
