+++ clients/python/setup.py
@@ -6,14 +6,14 @@ with open("README.md", "r", encoding="utf-8") as fh:
 setuptools.setup(
     name="stencil-python-client",
     version="0.0.1",
-    author="ODPF",
-    author_email="odpf@gmail.com",
+    author="Raystack",
+    author_email="raystack@gmail.com",
     description="Stencil Python client package provides a store to lookup protobuf descriptors and options to keep the protobuf descriptors upto date.",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    url="https://github.com/odpf/stencil",
+    url="https://github.com/raystack/stencil",
     project_urls={
-        "Bug Tracker": "https://github.com/odpf/stencil/issues",
+        "Bug Tracker": "https://github.com/raystack/stencil/issues",
     },
     classifiers=[
         "Programming Language :: Python :: 3",
@@ -23,4 +23,9 @@ setuptools.setup(
     package_dir={"": "src"},
     packages=setuptools.find_packages(where="src"),
     python_requires=">=3.6",
+    install_requires=[
+        'protobuf',
+        'schedule',
+        'requests'
+    ]
 )
